+++ src/GafferScene/SceneNode.cpp
@@ -41,7 +41,11 @@
 
 #include "boost/bind.hpp"
 
+#include "tbb/blocked_range.h"
+#include "tbb/parallel_reduce.h"
+
 using namespace std;
+using namespace tbb;
 using namespace Imath;
 using namespace IECore;
 using namespace GafferScene;
@@ -112,9 +116,19 @@ void SceneNode::affects( const Gaffer::Plug *input, AffectedPlugsContainer &outp
 			{
 				outputs.push_back( scenePlug->sortedChildNamesPlug() );
 			}
-			else if( input == scenePlug->sortedChildNamesPlug() )
+
+			if( input == scenePlug->sortedChildNamesPlug() )
 			{
 				outputs.push_back( scenePlug->existsPlug() );
+			}
+
+			if(
+				input == scenePlug->childNamesPlug() ||
+				input == scenePlug->boundPlug() ||
+				input == scenePlug->transformPlug()
+			)
+			{
+				outputs.push_back( scenePlug->childBoundsPlug() );
 			}
 		}
 	}
@@ -203,6 +217,10 @@ void SceneNode::hash( const ValuePlug *output, const Context *context, IECore::M
 		else if( output == scenePlug->sortedChildNamesPlug() )
 		{
 			hashSortedChildNames( context, scenePlug, h );
+		}
+		else if( output == scenePlug->childBoundsPlug() )
+		{
+			hashChildBounds( context, scenePlug, h );
 		}
 	}
 	else
@@ -335,6 +353,10 @@ void SceneNode::compute( ValuePlug *output, const Context *context ) const
 			{
 				static_cast<InternedStringVectorDataPlug *>( output )->setValue( computeSortedChildNames( context, scenePlug ) );
 			}
+			else if( output == scenePlug->childBoundsPlug() )
+			{
+				static_cast<AtomicBox3fPlug *>( output )->setValue( computeChildBounds( context, scenePlug ) );
+			}
 		}
 		else
 		{
@@ -384,66 +406,42 @@ IECore::ConstPathMatcherDataPtr SceneNode::computeSet( const IECore::InternedStr
 	throw IECore::NotImplementedException( string( typeName() ) + "::computeSet" );
 }
 
-IECore::MurmurHash SceneNode::hashOfTransformedChildBounds( const ScenePath &path, const ScenePlug *out, const IECore::InternedStringVectorData *childNamesData ) const
+Gaffer::ValuePlug::CachePolicy SceneNode::hashCachePolicy( const Gaffer::ValuePlug *output ) const
 {
-	ConstInternedStringVectorDataPtr computedChildNames;
-	if( !childNamesData )
+	if( auto parent = output->parent<ScenePlug>() )
 	{
-		computedChildNames = out->childNames( path );
-		childNamesData = computedChildNames.get();
-	}
-	const vector<InternedString> &childNames = childNamesData->readable();
-
-	IECore::MurmurHash result;
-	if( childNames.size() )
-	{
-		ScenePlug::PathScope pathScope( Context::current() );
-
-		ScenePath childPath( path );
-		childPath.push_back( InternedString() ); // room for the child name
-		for( vector<InternedString>::const_iterator it = childNames.begin(); it != childNames.end(); it++ )
+		if( output == parent->childBoundsPlug() )
 		{
-			childPath[path.size()] = *it;
-			pathScope.setPath( childPath );
-			out->boundPlug()->hash( result );
-			out->transformPlug()->hash( result );
+			return ValuePlug::CachePolicy::TaskIsolation;
 		}
 	}
-	else
+
+	return ComputeNode::hashCachePolicy( output );
+}
+
+Gaffer::ValuePlug::CachePolicy SceneNode::computeCachePolicy( const Gaffer::ValuePlug *output ) const
+{
+	if( auto parent = output->parent<ScenePlug>() )
 	{
-		result.append( typeId() );
-		result.append( "emptyBound" );
+		if( output == parent->childBoundsPlug() )
+		{
+			return ValuePlug::CachePolicy::TaskIsolation;
+		}
 	}
-	return result;
+
+	return ComputeNode::computeCachePolicy( output );
+}
+
+IECore::MurmurHash SceneNode::hashOfTransformedChildBounds( const ScenePath &path, const ScenePlug *out, const IECore::InternedStringVectorData *childNamesData ) const
+{
+	ScenePlug::PathScope pathScope( Context::current(), path );
+	return out->childBoundsPlug()->hash();
 }
 
 Imath::Box3f SceneNode::unionOfTransformedChildBounds( const ScenePath &path, const ScenePlug *out, const IECore::InternedStringVectorData *childNamesData ) const
 {
-	ConstInternedStringVectorDataPtr computedChildNames;
-	if( !childNamesData )
-	{
-		computedChildNames = out->childNames( path );
-		childNamesData = computedChildNames.get();
-	}
-	const vector<InternedString> &childNames = childNamesData->readable();
-
-	Box3f result;
-	if( childNames.size() )
-	{
-		ScenePlug::PathScope pathScope( Context::current() );
-
-		ScenePath childPath( path );
-		childPath.push_back( InternedString() ); // room for the child name
-		for( vector<InternedString>::const_iterator it = childNames.begin(); it != childNames.end(); it++ )
-		{
-			childPath[path.size()] = *it;
-			pathScope.setPath( childPath );
-			Box3f childBound = out->boundPlug()->getValue();
-			childBound = transform( childBound, out->transformPlug()->getValue() );
-			result.extendBy( childBound );
-		}
-	}
-	return result;
+	ScenePlug::PathScope pathScope( Context::current(), path );
+	return out->childBoundsPlug()->getValue();
 }
 
 void SceneNode::plugInputChanged( Gaffer::Plug *plug )
@@ -541,4 +539,95 @@ IECore::ConstInternedStringVectorDataPtr SceneNode::computeSortedChildNames( con
 	sorted->writable() = childNamesData->readable();
 	std::sort( sorted->writable().begin(), sorted->writable().end() );
 	return sorted;
+}
+
+void SceneNode::hashChildBounds( const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
+{
+	ComputeNode::hash( parent->childBoundsPlug(), context, h );
+	ConstInternedStringVectorDataPtr childNamesData = parent->childNamesPlug()->getValue();
+	const vector<InternedString> &childNames = childNamesData->readable();
+	if( childNames.empty() )
+	{
+		return;
+	}
+
+	const ThreadState &threadState = ThreadState::current();
+	using Range = blocked_range<size_t>;
+	tbb::task_group_context taskGroupContext( tbb::task_group_context::isolated );
+
+	h = parallel_deterministic_reduce(
+		Range( 0, childNames.size() ),
+		h,
+		[&] ( const Range &range, const MurmurHash &hash ) {
+
+			ScenePlug::PathScope pathScope( threadState );
+			auto childPath = context->get<ScenePath>( ScenePlug::scenePathContextName );
+			childPath.push_back( InternedString() ); // room for the child name
+
+			MurmurHash result = hash;
+			for( size_t i = range.begin(); i != range.end(); ++i )
+			{
+				childPath.back() = childNames[i];
+				pathScope.setPath( childPath );
+				parent->boundPlug()->hash( result );
+				parent->transformPlug()->hash( result );
+			}
+			return result;
+
+		},
+		[] ( const MurmurHash &x, const MurmurHash &y ) {
+
+			MurmurHash result = x;
+			result.append( y );
+			return result;
+		},
+		simple_partitioner(),
+		taskGroupContext
+	);
+}
+
+Imath::Box3f SceneNode::computeChildBounds( const Gaffer::Context *context, const ScenePlug *parent ) const
+{
+	ConstInternedStringVectorDataPtr childNamesData = parent->childNamesPlug()->getValue();
+	const vector<InternedString> &childNames = childNamesData->readable();
+	if( childNames.empty() )
+	{
+		return Box3f();
+	}
+
+	const ThreadState &threadState = ThreadState::current();
+	using Range = blocked_range<size_t>;
+	tbb::task_group_context taskGroupContext( tbb::task_group_context::isolated );
+
+	return tbb::parallel_reduce(
+		Range( 0, childNames.size() ),
+		Box3f(),
+		[&] ( const Range &range, const Box3f &bound ) {
+
+			ScenePlug::PathScope pathScope( threadState );
+			auto childPath = context->get<ScenePath>( ScenePlug::scenePathContextName );
+			childPath.push_back( InternedString() ); // room for the child name
+
+			Box3f result = bound;
+			for( size_t i = range.begin(); i != range.end(); ++i )
+			{
+				childPath.back() = childNames[i];
+				pathScope.setPath( childPath );
+				Box3f childBound = parent->boundPlug()->getValue();
+				childBound = transform( childBound, parent->transformPlug()->getValue() );
+				result.extendBy( childBound );
+			}
+			return result;
+
+		},
+		[] ( const Box3f &x, const Box3f &y ) {
+
+			Box3f result = x;
+			result.extendBy( y );
+			return result;
+
+		},
+		tbb::auto_partitioner(),
+		taskGroupContext
+	);
 }