+++ Changes.md
@@ -4,6 +4,7 @@
 Improvements
 ------------
 
+- SceneNode : Improved performance for all nodes that must propagate bounds from children to parents.
 - PointsType : Removed unnecessary bounds computation overhead.
 
 Breaking Changes